Moving out of decision_review_v1, this is a v0 matter Moving code up a layer as to not affect other people using FormAttachmentCreate class
